+++ src/main/java/baseball/Domain/BaseballScore.java
@@ -1,4 +1,4 @@
-package baseball;
+package baseball.Domain;
 
 public class BaseballScore {
