+++ src/test/java/baseball/BaseballScoreTest.java
@@ -1,5 +1,6 @@
 package baseball;
 
+import baseball.Domain.BaseballScore;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
